Create main.py
This file include the Graphic User Interface but not the required LLM pre-trained model access, the call for the financial market info (stocks, index and commodities) nor the AI search_engine. Can rub by itself but wont create any file, nor a financial report.


To access both the pre-trained LLM or the AI-search engine, please contact any of the authors.